Migrate files after refactor
