+++ services/agents/src/contracts/component-design.ts
